refactor: move views logic to separate directory
