+++ ui/src/dashboards/components/EditVEO.tsx
@@ -12,7 +12,7 @@ import VEOHeader from 'src/dashboards/components/VEOHeader'
 // Actions
 import {setName} from 'src/timeMachine/actions'
 import {saveVEOView} from 'src/dashboards/actions/thunks'
-import {getViewForTimeMachine} from 'src/dashboards/actions/views'
+import {getViewForTimeMachine} from 'src/views/actions/thunks'
 
 // Utils
 import {getActiveTimeMachine} from 'src/timeMachine/selectors'
