+++ ui/src/dashboards/components/dashboard_index/DashboardCard.tsx
@@ -17,6 +17,7 @@ import {
   removeDashboardLabel,
 } from 'src/dashboards/actions/thunks'
 import {createLabel as createLabelAsync} from 'src/labels/actions'
+import {resetViews} from 'src/views/actions/creators'
 
 // Selectors
 import {viewableLabels} from 'src/labels/selectors'
@@ -26,7 +27,6 @@ import {AppState, Label} from 'src/types'
 
 // Constants
 import {DEFAULT_DASHBOARD_NAME} from 'src/dashboards/constants'
-import {resetViews} from 'src/dashboards/actions/views'
 
 // Utilities
 import {relativeTimestampFormatter} from 'src/shared/utils/relativeTimestampFormatter'
