+++ ui/src/dashboards/selectors/index.ts
@@ -1,15 +1,6 @@
 import {get} from 'lodash'
 
-import {
-  AppState,
-  View,
-  Check,
-  ViewType,
-  RemoteDataState,
-  TimeRange,
-  ResourceType,
-  Dashboard,
-} from 'src/types'
+import {AppState, View, Check, ViewType, TimeRange} from 'src/types'
 
 import {
   getValuesForVariable,
@@ -19,15 +10,6 @@ import {
 
 // Constants
 import {DEFAULT_TIME_RANGE} from 'src/shared/constants/timeRanges'
-import {getByID} from 'src/resources/selectors'
-
-export const getView = (state: AppState, id: string): View => {
-  return get(state, `views.views.${id}.view`)
-}
-
-export const getViewStatus = (state: AppState, id: string): RemoteDataState => {
-  return get(state, `views.views.${id}.status`, RemoteDataState.Loading)
-}
 
 export const getTimeRangeByDashboardID = (
   state: AppState,
@@ -44,25 +26,6 @@ export const getCheckForView = (
   return viewType === 'check'
     ? state.checks.list.find(c => c.id === checkID)
     : null
-}
-
-export const getViewsForDashboard = (
-  state: AppState,
-  dashboardID: string
-): View[] => {
-  const dashboard = getByID<Dashboard>(
-    state,
-    ResourceType.Dashboards,
-    dashboardID
-  )
-
-  const cellIDs = new Set(dashboard.cells.map(cellID => cellID))
-
-  const views = Object.values(state.views.views)
-    .map(d => d.view)
-    .filter(view => view && cellIDs.has(view.cellID))
-
-  return views
 }
 
 interface DropdownValues {
